refactor : 함수명 및 파일 경로 변경
